Restructure node-template so it is clear that node, runtime, and pallets are separated
