+++ bin/node-template/Cargo.toml
@@ -2,13 +2,13 @@
 name = "node-template"
 version = "2.0.0"
 authors = ["Anonymous"]
-build = "build.rs"
 edition = "2018"
 license = "Unlicense"
+build = "node/build.rs"
 
 [[bin]]
 name = "node-template"
-path = "src/main.rs"
+path = "node/main.rs"
 
 [dependencies]
 futures = "0.3.1"
@@ -27,7 +27,7 @@ sp-consensus = { version = "0.8", path = "../../primitives/consensus/common" }
 grandpa = { version = "0.8", package = "sc-finality-grandpa", path = "../../client/finality-grandpa" }
 grandpa-primitives = { version = "2.0.0", package = "sp-finality-grandpa", path = "../../primitives/finality-grandpa" }
 sc-client = { version = "0.8", path = "../../client/" }
-node-template-runtime = { version = "2.0.0", path = "runtime" }
+node-template-runtime = { version = "2.0.0", path = "node/runtime" }
 sp-runtime = { version = "2.0.0", path = "../../primitives/runtime" }
 sc-basic-authorship = { path = "../../client/basic-authorship" }
 structopt = "0.3.8"
