+++ bin/node-template/node/cli.rs
