+++ bin/node-template/node/command.rs
